Change the backend, add JSON endpoints
