+++ src/main.rs
@@ -1,6 +1,7 @@
 #![feature(proc_macro_hygiene, decl_macro, custom_attribute)]
 
 mod id_map;
+mod error;
 mod store;
 
 #[macro_use]
@@ -17,9 +18,12 @@ use std::path::{Path, PathBuf};
 use maud::{html, Markup, DOCTYPE};
 use rocket::response::NamedFile;
 use rocket_contrib::serve::StaticFiles;
+use rocket_contrib::json::Json;
+use rocket_contrib::json::JsonValue;
 use chrono::prelude::*;
 
-use store::Store;
+use store::{Store, Id, Event, action::Actions};
+use error::*;
 
 #[get("/")]
 fn index(store: Store) -> Markup {
@@ -76,11 +80,42 @@ fn admin_route() -> Option<NamedFile> {
     admin()
 }
 
-#[get("/admin/<path..>")]
+#[post("/admin/event", data = "<obj>")]
+fn store_event(store: Store, obj: Json<Event>) -> Result<String> {
+    store.create(obj.0)
+        .map_err(|x| Error::Database(x.to_string()))
+        .map(|x| x.to_string())
+}
+
+#[post("/admin/event/<id>", data = "<obj>")]
+fn update_event(store: Store, id: String, obj: Json<Event>) -> Result<Json<Event>> {
+    Id::parse_str(&id)
+        .map_err(|x| Error::ParseId(x.to_string()))
+        .and_then(|x| store.update(x, obj.0).map_err(|x| Error::Database(x.to_string())))
+        .map(|x| Json(x))
+}
+
+#[get("/admin/event/<id>")]
+fn get_event(store: Store, id: String) -> Result<Json<Event>> {
+    Id::parse_str(&id)
+        .map_err(|x| Error::ParseId(x.to_string()))
+        .and_then(|x| store.read(x).map_err(|x| Error::Database(x.to_string())))
+        .map(|x| Json(x))
+}
+
+#[get("/admin/event/<id>/delete")]
+fn delete_event(store: Store, id: String) -> Result<Json<Event>> {
+    Id::parse_str(&id)
+        .map_err(|x| Error::ParseId(x.to_string()))
+        .and_then(|x| store.delete(x).map_err(|x| Error::Database(x.to_string())))
+        .map(|x| Json(x))
+}
+
+/*#[get("/admin/<path..>")]
 #[allow(unused_variables)]
 fn admin_subroute(path: PathBuf) -> Option<NamedFile> {
     admin()
-}
+}*/
 
 fn admin() -> Option<NamedFile> {
     NamedFile::open(Path::new("admin/dist/index.html")).ok()
@@ -93,6 +128,6 @@ fn main() {
             "/static",
             StaticFiles::from(concat!(env!("CARGO_MANIFEST_DIR"), "/static")),
         )
-        .mount("/", routes![index, admin_route, admin_subroute,])
+        .mount("/", routes![index, admin_route, get_event, store_event, update_event, delete_event])
         .launch();
 }
